Rework jb_run to be more reliable. Before, we were setting the stdout redirection pipe to nonblocking, allowing read() calls to fail with EWOULDBLOCK/EAGAIN. However, this could be unreliable: if we don't read the pipe fast enough, we could lose output from the child process.
This change keeps the pipe in blocking-mode. We then periodically check if we can read the pipe without blocking using poll().

This change also cleans up the duplication of work between jb_run and jb_run_get_output by abstracting away the main functionality to an internal function that accepts a call back to use for marshalling child process output.